sentify/requirements.txt
pdfminer.six
smart_open
Wikipedia-API
twine
pysbd
python-docx
openpyxl
pandas